+++ puma-admin/src/main/java/com/dianping/puma/admin/web/SrcDBInstanceController.java
@@ -85,7 +85,7 @@ public class SrcDBInstanceController {
 	public String createPost(
 			String id,
 			String name,
-			Integer serverId,
+			Long serverId,
 			String ip,
 			String username,
 			String password) {
